requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
#dag-factory>=0.23.0a7
apache-airflow-providers-postgres
astronomer-cosmos
faker